fix after review: changed how the gas model is used in tests
